Add required structs, Draw(), and RiemannMap().
ComplexRectangle is a rectangle formed by two complex numbers. There are
no public methods because they are not needed. If some are, then they
can be added.

ComplexMap is a complex function of one variable.

ColorMap is a function which maps a point in the complex plane onto an
RGB color (there will be predefined ColorMap converters, it is
recommended that those are used).

Draw creates an image given a ColorMap, the size of the image, and a
(rectangular) domain.

RiemannMap is a standard way of converting from a ComplexMap to
ColorMap. It works by converting the complex number to its Riemann
sphere coordinate tuple, then respectively uniformly mapping that tuple
to red, green, and blue. This may change depending on how it looks
aesthetically, as there is no real reason for uniform mapping except
simplicity.